fix:top components moved from route directories to component directories
