+++ src/app/_components/auth/Login.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import React, { useState } from "react";
-import { useAuthActions } from "../../../../hooks/user/useAuthActions";
+import { useAuthActions } from "../../../hooks/user/useAuthActions";
 import { Box, Button, Grid, TextField, Typography } from "@mui/material";
 
 export const Login = () => {
