+++ src/app/_components/userpage/UserPage.tsx
@@ -5,10 +5,10 @@ import { useRecoilValue } from "recoil";
 import Link from "next/link";
 import { useParams } from "next/navigation";
 import { loginUserState } from "../../../store/atoms/userAtom";
-import { ImageItem } from "../../_components/userpage/ImageItem";
-import { FollowModal } from "../../_components/userpage/FollowModal";
-import { FollowButton } from "../../_components/FollowButton";
-import { ImageSortSelector } from "../../_components/userpage/ImageSortSelector";
+import { ImageItem } from "./ImageItem";
+import { FollowModal } from "./FollowModal";
+import { FollowButton } from "../FollowButton";
+import { ImageSortSelector } from "./ImageSortSelector";
 import { useUsers } from "../../../hooks/user/useUsers";
 import { useImages } from "../../../hooks/useImages";
 import { useParks } from "../../../hooks/useParks";
